json marshaling extenstion
This extension is not mean to be used under heavy load, as it
runs about 20% slower than encoding/json, and lacking some
importnt facilities like "omitempty".

You are encouraged to write json tags on your own. But if you are
just writing a simple/small app and you are lazy, it is here for
you.